+++ inst/assets/buttons-group/checkboxGroupButtons-bindings.js
@@ -129,8 +129,10 @@ $.extend(checkboxGroupButtonsBinding, {
 
     if (data.disabled) {
       $el.find("button").attr("disabled", "disabled");
+      $el.find("button").addClass("disabled");
     } else {
       $el.find("button").removeAttr("disabled");
+      $el.find("button").removeClass("disabled");
     }
     if (data.hasOwnProperty("disabledChoices")) {
       for (var i = 0; i < data.disabledChoices.length; i++) {
@@ -142,7 +144,8 @@ $.extend(checkboxGroupButtonsBinding, {
             '"]'
         )
           .parent()
-          .attr("disabled", "disabled");
+          .attr("disabled", "disabled")
+          .addClass("disabled");
       }
     }